+++ packages/types/test/runtime-test.types.ts
@@ -63,6 +63,8 @@ test("matchers are typed against the value", () => {
   expect([1, 2]).toEqual([1, 2]);
   expect({ answer: 42 }).toMatchSnapshot();
   expect({ answer: 42 }).toMatchSnapshot("answer");
+  expect({ id: 42 }).toMatchSnapshot({ id: expect.any(Number) });
+  expect("<main />").toMatchFileSnapshot("home.html");
   assertSnapshot({ answer: 42 }, "answer assertion");
   expect(() => { throw new Error("broken"); }).toThrowErrorMatchingSnapshot();
   expect({ a: 1 }).toMatchObject({ a: 1 });